Completed Etch-a-Sketch
Created a function that will create the grid.
I created a button that ask the user's input for number of rows and columns.
When  user enters number it will call the function I had created
and create a new grid with user's input value.
When the user enters the input, it will remove the previous grid.